# apps/web/src/app/(main)/styleIST/arab-stylist-studio-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Arab Stylist Studio API
  version: 0.1.0
  summary: واجهات API الأساسية لمنصة Stylist عربية متعددة المستخدمين
  description: |
    هذه المواصفة تمثل نقطة بداية تنفيذية للنسخة الأولى من المنتج.
    جميع العمليات الحساسة تعتمد على مصادقة خادمية وصلاحيات مرتبطة بالـ tenant والمشروع.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
security:
  - sessionAuth: []
  - bearerAuth: []
...
paths:
...
  /productions/{productionId}/characters:
    get:
      tags: [Characters]
      summary: قائمة الشخصيات
      operationId: listCharacters
      parameters:
        - $ref: "#/components/parameters/ProductionId"
      responses:
        "200":
          description: Characters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterListResponse"
    post:
      tags: [Characters]
      summary: إنشاء شخصية
      operationId: createCharacter
      parameters:
        - $ref: "#/components/parameters/ProductionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterRequest"
      responses:
        "201":
          description: Character created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
  /characters/{characterId}:
    get:
      tags: [Characters]
      summary: قراءة شخصية
      operationId: getCharacter
      parameters:
        - $ref: "#/components/parameters/CharacterId"
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
    patch:
      tags: [Characters]
      summary: تعديل شخصية
      operationId: updateCharacter
      parameters:
        - $ref: "#/components/parameters/CharacterId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterRequest"
      responses:
        "200":
          description: Character updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
...
  /vendors:
    get:
      tags: [Vendors]
      summary: قائمة الموردين
      operationId: listVendors
      parameters:
        - name: production_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
    post:
      tags: [Vendors]
      summary: إنشاء مورد
      operationId: createVendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorResponse"
...
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: حالة Job
      operationId: getJobStatus
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProductionId:
      {
        name: productionId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    CharacterId:
      {
        name: characterId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    JobId:
      {
        name: jobId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
  schemas:
    UUID: { type: string, format: uuid }
    Meta:
      type: object
      properties:
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        retryable: { type: boolean }
    Envelope:
      type: object
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
        error:
          oneOf:
            - { $ref: "#/components/schemas/Error" }
            - { type: "null" }
...
    RunStatus:
      { type: string, enum: [queued, running, succeeded, failed, cancelled] }
...
    Character:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        production_id: { $ref: "#/components/schemas/UUID" }
        canonical_name: { type: string }
        aliases_jsonb: { type: array, items: {} }
        gender_presentation: { type: string, nullable: true }
        age_band: { type: string, nullable: true }
        social_class_notes: { type: string, nullable: true }
        hometown_notes: { type: string, nullable: true }
        narrative_role: { type: string, nullable: true }
        style_status: { type: string }
...
    Vendor:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        region: { type: string, nullable: true }
        catalog_url: { type: string, nullable: true }
...
    CreateCharacterRequest:
      type: object
      required: [canonical_name]
      properties:
        canonical_name: { type: string }
        aliases_jsonb: { type: array, items: {} }
        gender_presentation: { type: string, nullable: true }
        age_band: { type: string, nullable: true }
        social_class_notes: { type: string, nullable: true }
        hometown_notes: { type: string, nullable: true }
        narrative_role: { type: string, nullable: true }
    UpdateCharacterRequest:
      type: object
      properties:
        canonical_name: { type: string }
        aliases_jsonb: { type: array, items: {} }
        gender_presentation: { type: string, nullable: true }
        age_band: { type: string, nullable: true }
        social_class_notes: { type: string, nullable: true }
        hometown_notes: { type: string, nullable: true }
        narrative_role: { type: string, nullable: true }
        style_status: { type: string }
    CharacterResponse:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Character" }
    CharacterListResponse:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  {
                    type: array,
                    items: { $ref: "#/components/schemas/Character" },
                  }
...
    CreateVendorRequest:
      type: object
      required: [name]
      properties:
        production_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        region: { type: string, nullable: true }
        catalog_url: { type: string, nullable: true }
        contact_jsonb: { type: object, additionalProperties: true }
    VendorResponse:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Vendor" }
    VendorListResponse:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  {
                    type: array,
                    items: { $ref: "#/components/schemas/Vendor" },
                  }
...
    JobStatusResponse:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                id: { $ref: "#/components/schemas/UUID" }
                status: { $ref: "#/components/schemas/RunStatus" }
                progress: { type: number, nullable: true }
                result_ref:
                  { type: object, nullable: true, additionalProperties: true }
                error:
                  { type: object, nullable: true, additionalProperties: true }
...